requirements.txt
aiofiles==0.6.0
keyboard==0.13.5
python-dotenv==0.15.0
